# Optimize get blocks not confirmed sorted query