… safe_transaction_service/history/migrations/0092_remove_ethereumblock_history_block_confirmed_idx_and_more.py