- There made no sense to add `processed` to the index, as it's always `False`
- By adding number to the index sorting is already done

```
staging_transaction_service=> EXPLAIN ANALYZE SELECT history_ethereumblock.number, history_ethereumblock.gas_limit, history_ethereumblock.gas_used, history_ethereumblock.timestamp, history_ethereumblock.block_hash, history_ethereumblock.parent_hash, history_ethereumblock.confirmed
FROM history_ethereumblock
WHERE NOT history_ethereumblock.confirmed
ORDER BY history_ethereumblock.number ASC
LIMIT 10;
                                                                              QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=1750.05..1750.07 rows=10 width=90) (actual time=5.244..5.246 rows=10 loops=1)
   ->  Sort  (cost=1750.05..1751.08 rows=413 width=90) (actual time=5.243..5.244 rows=10 loops=1)
         Sort Key: number
         Sort Method: top-N heapsort  Memory: 26kB
         ->  Index Scan using history_block_confirmed_idx on history_ethereumblock  (cost=0.27..1741.12 rows=413 width=90) (actual time=5.084..5.221 rows=89 loops=1)
 Planning Time: 0.334 ms
 Execution Time: 5.304 ms
(7 rows)

staging_transaction_service=> CREATE INDEX test on history_ethereumblock(number) WHERE confirmed = False;
CREATE INDEX
staging_transaction_service=> EXPLAIN ANALYZE SELECT history_ethereumblock.number, history_ethereumblock.gas_limit, history_ethereumblock.gas_used, history_ethereumblock.timestamp, history_ethereumblock.block_hash, history_ethereumblock.parent_hash, history_ethereumblock.confirmed
FROM history_ethereumblock
WHERE NOT history_ethereumblock.confirmed
ORDER BY history_ethereumblock.number ASC
LIMIT 10;
                                                              QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.14..0.73 rows=10 width=90) (actual time=0.012..0.016 rows=10 loops=1)
   ->  Index Scan using test on history_ethereumblock  (cost=0.14..24.30 rows=413 width=90) (actual time=0.011..0.014 rows=10 loops=1)
 Planning Time: 0.158 ms
 Execution Time: 0.050 ms
(4 rows)

staging_transaction_service=> EXPLAIN ANALYZE SELECT history_ethereumblock.number, history_ethereumblock.gas_limit, history_ethereumblock.gas_used, history_ethereumblock.timestamp, history_ethereumblock.block_hash, history_ethereumblock.parent_hash, history_ethereumblock.confirmed
FROM history_ethereumblock
WHERE NOT history_ethereumblock.confirmed
ORDER BY history_ethereumblock.number ASC
LIMIT 10;
                                                              QUERY PLAN
---------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.14..0.73 rows=10 width=90) (actual time=0.025..0.029 rows=10 loops=1)
   ->  Index Scan using test on history_ethereumblock  (cost=0.14..24.30 rows=413 width=90) (actual time=0.025..0.028 rows=10 loops=1)
 Planning Time: 0.085 ms
 Execution Time: 0.062 ms
(4 rows)
```